refactor general settings
