+++ gui/src/inventory/pearSettings/components/LoadingPlaceholder.tsx
@@ -1,0 +1,3 @@
+export const LoadingPlaceholder = () => (
+  <div className="animate-pulse bg-list-hoverBackground h-6 w-16 rounded"></div>
+);
